Refactor a bit and test checkpoitner cleanup
